chore: extract PeriodTypeStore
