+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/period/BiMonthlyPeriodType.java
@@ -32,6 +32,7 @@ package org.hisp.dhis.period;
 import com.google.common.collect.Lists;
 import java.util.Date;
 import java.util.List;
+import javax.annotation.Nonnull;
 import org.hisp.dhis.calendar.Calendar;
 import org.hisp.dhis.calendar.DateTimeUnit;
 
@@ -139,6 +140,7 @@ public class BiMonthlyPeriodType extends CalendarPeriodType {
     return String.format("%d%02dB", dateTimeUnit.getYear(), (dateTimeUnit.getMonth() + 1) / 2);
   }
 
+  @Nonnull
   @Override
   public String getIsoFormat() {
     return ISO_FORMAT;
@@ -157,6 +159,7 @@ public class BiMonthlyPeriodType extends CalendarPeriodType {
     return cal.toIso(dateTimeUnit).toJdkDate();
   }
 
+  @Nonnull
   @Override
   public String getIso8601Duration() {
     return ISO8601_DURATION;
