+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/period/BiWeeklyAbstractPeriodType.java
@@ -36,6 +36,7 @@ import java.time.temporal.WeekFields;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
+import javax.annotation.Nonnull;
 import org.hisp.dhis.calendar.Calendar;
 import org.hisp.dhis.calendar.DateTimeUnit;
 
@@ -78,6 +79,7 @@ public abstract class BiWeeklyAbstractPeriodType extends CalendarPeriodType {
     return startOfWeek;
   }
 
+  @Nonnull
   @Override
   public String getIso8601Duration() {
     return isoDuration;
@@ -93,6 +95,7 @@ public abstract class BiWeeklyAbstractPeriodType extends CalendarPeriodType {
     return sqlInterval;
   }
 
+  @Nonnull
   @Override
   public String getIsoFormat() {
     return isoFormat;
