+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/period/FinancialNovemberPeriodType.java
@@ -30,6 +30,7 @@
 package org.hisp.dhis.period;
 
 import java.util.Calendar;
+import javax.annotation.Nonnull;
 import org.hisp.dhis.calendar.DateTimeUnit;
 
 /**
@@ -58,11 +59,13 @@ public class FinancialNovemberPeriodType extends FinancialPeriodType {
     return String.format("%dNov", dateTimeUnit.getYear() + 1);
   }
 
+  @Nonnull
   @Override
   public String getIsoFormat() {
     return ISO_FORMAT;
   }
 
+  @Nonnull
   @Override
   public String getIso8601Duration() {
     return ISO8601_DURATION;
