+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/period/FinancialSeptemberPeriodType.java
@@ -30,6 +30,7 @@
 package org.hisp.dhis.period;
 
 import java.util.Calendar;
+import javax.annotation.Nonnull;
 import org.hisp.dhis.calendar.DateTimeUnit;
 
 /**
@@ -58,11 +59,13 @@ public class FinancialSeptemberPeriodType extends FinancialPeriodType {
     return String.format("%dSep", dateTimeUnit.getYear());
   }
 
+  @Nonnull
   @Override
   public String getIsoFormat() {
     return ISO_FORMAT;
   }
 
+  @Nonnull
   @Override
   public String getIso8601Duration() {
     return ISO8601_DURATION;
