+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/period/MonthlyPeriodType.java
@@ -32,6 +32,7 @@ package org.hisp.dhis.period;
 import com.google.common.collect.Lists;
 import java.util.Date;
 import java.util.List;
+import javax.annotation.Nonnull;
 import org.hisp.dhis.calendar.Calendar;
 import org.hisp.dhis.calendar.DateTimeUnit;
 
@@ -135,11 +136,13 @@ public class MonthlyPeriodType extends CalendarPeriodType {
     return String.format("%d%02d", dateTimeUnit.getYear(), dateTimeUnit.getMonth());
   }
 
+  @Nonnull
   @Override
   public String getIsoFormat() {
     return ISO_FORMAT;
   }
 
+  @Nonnull
   @Override
   public String getIso8601Duration() {
     return ISO8601_DURATION;
