+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/config/StartupConfig.java
@@ -37,8 +37,8 @@ import org.hisp.dhis.datavalue.DataValueChangelogStore;
 import org.hisp.dhis.external.conf.DhisConfigurationProvider;
 import org.hisp.dhis.i18n.I18nLocaleService;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
-import org.hisp.dhis.period.PeriodStore;
 import org.hisp.dhis.period.PeriodTypePopulator;
+import org.hisp.dhis.period.PeriodTypeStore;
 import org.hisp.dhis.scheduling.JobScheduler;
 import org.hisp.dhis.setting.SystemSettingsService;
 import org.hisp.dhis.startup.ConfigurationPopulator;
@@ -58,8 +58,8 @@ import org.springframework.context.annotation.Configuration;
 @Configuration
 public class StartupConfig {
   @Bean("org.hisp.dhis.period.PeriodTypePopulator")
-  public PeriodTypePopulator periodTypePopulator(PeriodStore periodStore) {
-    PeriodTypePopulator populator = new PeriodTypePopulator(periodStore);
+  public PeriodTypePopulator periodTypePopulator(PeriodTypeStore periodTypeStore) {
+    PeriodTypePopulator populator = new PeriodTypePopulator(periodTypeStore);
     populator.setName("PeriodTypePopulator");
     populator.setRunlevel(3);
     return populator;
