+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/period/PeriodTypePopulator.java
@@ -40,10 +40,10 @@ import org.hisp.dhis.system.startup.TransactionContextStartupRoutine;
 @RequiredArgsConstructor
 public class PeriodTypePopulator extends TransactionContextStartupRoutine {
 
-  private final PeriodStore periodStore;
+  private final PeriodTypeStore periodTypeStore;
 
   @Override
   public void executeInTransaction() {
-    PeriodType.getAvailablePeriodTypes().forEach(periodStore::addPeriodType);
+    PeriodType.getAvailablePeriodTypes().forEach(periodTypeStore::addPeriodType);
   }
 }
