+++ dhis-2/dhis-support/dhis-support-db-migration/src/main/resources/org/hisp/dhis/db/migration/2.44/V2_44_19__add_periodtype_labels.sql
@@ -1,1 +1,7 @@
-alter table periodtype add column if not exists translations jsonb not null default '[]';
+ALTER TABLE periodtype ADD COLUMN IF NOT EXISTS translations jsonb NOT NULL DEFAULT '[]';
+
+CREATE TABLE relativeperiod (
+  name character varying(50) PRIMARY KEY,
+  label character varying(255),
+  translations jsonb NOT NULL DEFAULT '[]'
+);
