+++ dhis-2/dhis-support/dhis-support-hibernate/src/main/java/org/hisp/dhis/hibernate/HibernateAutoTransactionStore.java
@@ -1,0 +1,95 @@
+/*
+ * Copyright (c) 2004-2026, University of Oslo
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ * list of conditions and the following disclaimer.
+ *
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ * this list of conditions and the following disclaimer in the documentation
+ * and/or other materials provided with the distribution.
+ *
+ * 3. Neither the name of the copyright holder nor the names of its contributors 
+ * may be used to endorse or promote products derived from this software without
+ * specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
+ * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package org.hisp.dhis.hibernate;
+
+import jakarta.persistence.EntityManager;
+import java.sql.Connection;
+import java.util.function.Function;
+import javax.sql.DataSource;
+import lombok.extern.slf4j.Slf4j;
+import org.hibernate.StatelessSession;
+import org.hibernate.Transaction;
+import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.datasource.DataSourceUtils;
+import org.springframework.transaction.support.TransactionSynchronizationManager;
+
+@Slf4j
+public abstract class HibernateAutoTransactionStore<T> extends HibernateGenericStore<T> {
+
+  private final DataSource dataSource;
+
+  public HibernateAutoTransactionStore(
+      EntityManager entityManager,
+      DataSource dataSource,
+      JdbcTemplate jdbcTemplate,
+      ApplicationEventPublisher publisher,
+      Class<T> clazz,
+      boolean cacheable) {
+    super(entityManager, jdbcTemplate, publisher, clazz, cacheable);
+    this.dataSource = dataSource;
+  }
+
+  protected final <R> R runAutoJoinTransaction(Function<StatelessSession, R> query) {
+    boolean active = TransactionSynchronizationManager.isActualTransactionActive();
+    boolean readOnly = TransactionSynchronizationManager.isCurrentTransactionReadOnly();
+
+    if (active && !readOnly) {
+      // run in existing TX (for visibility)
+      Connection borrowedConnection = DataSourceUtils.getConnection(dataSource);
+      StatelessSession session =
+          getSession().getSessionFactory().openStatelessSession(borrowedConnection);
+      return query.apply(session);
+    }
+
+    // run in new TX
+    StatelessSession session = getSession().getSessionFactory().openStatelessSession();
+
+    Transaction transaction = null;
+    try {
+      transaction = session.beginTransaction();
+      R result = query.apply(session);
+      transaction.commit();
+      return result;
+    } catch (RuntimeException e) {
+      // Handle rollback for self-managed transactions
+      if (transaction != null && transaction.isActive()) {
+        transaction.rollback();
+      }
+      throw e;
+    } finally {
+      try {
+        session.close();
+      } catch (Exception e) {
+        log.error("Session close error", e);
+      }
+    }
+  }
+}
