+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/period/PeriodStoreTest.java
@@ -49,13 +49,14 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional
 class PeriodStoreTest extends PostgresIntegrationTestBase {
   @Autowired private PeriodStore periodStore;
+  @Autowired private PeriodTypeStore periodTypeStore;
 
   // -------------------------------------------------------------------------
   // Period
   // -------------------------------------------------------------------------
   @Test
   void testAddPeriod() {
-    List<PeriodType> periodTypes = periodStore.getAllPeriodTypes();
+    List<PeriodType> periodTypes = periodTypeStore.getAllPeriodTypes();
     Iterator<PeriodType> it = periodTypes.iterator();
     PeriodType periodTypeA = it.next();
     PeriodType periodTypeB = it.next();
@@ -91,7 +92,7 @@ class PeriodStoreTest extends PostgresIntegrationTestBase {
 
   @Test
   void testGetAllPeriods() {
-    List<PeriodType> periodTypes = periodStore.getAllPeriodTypes();
+    List<PeriodType> periodTypes = periodTypeStore.getAllPeriodTypes();
     Iterator<PeriodType> it = periodTypes.iterator();
     PeriodType periodTypeA = it.next();
     PeriodType periodTypeB = it.next();
